пакет выбранного сервера

+++ authserver/src/main/java/ru/jts_dev/authserver/packets/IncomingMessageWrapper.java
@@ -38,7 +38,7 @@ public abstract class IncomingMessageWrapper implements Message<ByteBuf>, Runnab
     public abstract void prepare();
 
 
-    public int readByte() {
+    public byte readByte() {
         if (buffer.readableBytes() < Byte.BYTES)
             throw new IndexOutOfBoundsException("At least 1 byte1 must be readable in buffer");
